Switch to rendering interface after initial API call (success + fail) and relying on cached shutdown state so that button text matches button action
I basically went the way Michael suggested in the end because users will be confused if they press "Join meeting" on an old tab and it comments saying they've started a meeting.

Also I stopped rendering on page load because the swapping out of the start/join button is janky - seems nicer to wait for API response so we render once and correct.